# docs/swagger.yaml
definitions:
  domain.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      role:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      uuid:
        type: string
    type: object
...
  dto.ErrorResponse:
    description: ErrorResponse provides a consistent error format.
    properties:
      error:
        type: string
    type: object
...
  dto.LoginRequest:
    description: LoginRequest validates input for user login. Email must be a valid
      email address. Password is required.
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    description: LoginResponse includes the authenticated user's details.
    properties:
      user:
        $ref: '#/definitions/domain.User'
    type: object
  dto.RegisterUserRequest:
    description: RegisterUserRequest validates input for user registration. FullName
      must be at least 3 and at max 255 characters long. Email must be a valid email
      address. Password must be at least 8 and at max 64 characters long.
    properties:
      email:
        type: string
      fullName:
        maxLength: 255
        minLength: 3
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    required:
    - email
    - fullName
    - password
    type: object
  dto.RegisterUserResponse:
    description: RegisterUserResponse includes the created user's details.
    properties:
      user:
        $ref: '#/definitions/domain.User'
    type: object
...
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: |-
        Verifies password using bcrypt comparison.
        Generates new JWT access token using ECDSA encryption.
        Sets HTTP-only cookie with new access token and and X-Request-Id header.
      parameters:
      - description: User login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Authenticate a user and provide access tokens
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: |-
        Hashes password using bcrypt before storage.
        Generates JWT access token using ECDSA encryption.
        Sets HTTP-only cookie with access token and X-Request-Id header.
      parameters:
      - description: User registration details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"